Mchihab

+++ libft/ft_calloc.c
@@ -3,20 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: smarsi <smarsi@student.42.fr>              +#+  +:+       +#+        */
+/*   By: mchihab <mchihab@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 15:40:59 by smarsi            #+#    #+#             */
-/*   Updated: 2024/02/25 09:52:04 by smarsi           ###   ########.fr       */
+/*   Updated: 2024/05/21 11:39:54 by mchihab          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_calloc(size_t count, size_t size)
+void	*ft_calloc(int count, int size)
 {
 	void	*ptr;
 
-	ptr = malloc(count * size);
+	ptr = malloc(count * size );
 	if (!ptr)
 		return (ptr);
 	ft_bzero(ptr, count * size);
